requirements.txt
grpcio-tools==1.74.0
grpcio>=1.71.2
litellm>=1.74.0
python-dotenv>=1.0.1
qdrant-client>=1.9.0
protobuf>=6.31.0
tiktoken==0.11.0
dspy-ai
huggingface_hub
numpy<2
tqdm>=4.66.0
openai>=1.90.0,<1.110.0
pydantic>=2.7.3
